# Conflicts:
#	x-pack/plugins/siem/public/containers/errors/index.tsx

+++ x-pack/plugins/code/server/lsp/abstract_launcher.ts
@@ -5,14 +5,21 @@
  */
 
 import { ChildProcess } from 'child_process';
+import { ResponseError } from 'vscode-jsonrpc';
 import { ILanguageServerLauncher } from './language_server_launcher';
 import { ServerOptions } from '../server_options';
 import { LoggerFactory } from '../utils/log_factory';
 import { Logger } from '../log';
 import { LanguageServerProxy } from './proxy';
 import { RequestExpander } from './request_expander';
+import { LanguageServerStartFailed } from '../../common/lsp_error_codes';
 
 let seqNo = 1;
+
+export const ServerStartFailed = new ResponseError(
+  LanguageServerStartFailed,
+  'Launch language server failed.'
+);
 
 export abstract class AbstractLauncher implements ILanguageServerLauncher {
   running: boolean = false;
@@ -20,7 +27,9 @@ export abstract class AbstractLauncher implements ILanguageServerLauncher {
   private child: ChildProcess | null = null;
   private _startTime: number = -1;
   private _proxyConnected: boolean = false;
-  private readonly log: Logger;
+  protected readonly log: Logger;
+  private spawnTimes: number = 0;
+  private launchReject?: ((reason?: any) => void);
   protected constructor(
     readonly name: string,
     readonly targetHost: string,
@@ -49,6 +58,7 @@ export abstract class AbstractLauncher implements ILanguageServerLauncher {
       });
     } else {
       child = await this.spawnProcess(installationPath, port, log);
+      this.spawnTimes += 1;
       this.child = child;
       log.debug('spawned a child process ' + child.pid);
       this._currentPid = child.pid;
@@ -79,11 +89,15 @@ export abstract class AbstractLauncher implements ILanguageServerLauncher {
     });
     proxy.listen();
     this.startConnect(proxy);
-    await new Promise(resolve => {
+    await new Promise((resolve, reject) => {
       proxy.onConnected(() => {
         this._proxyConnected = true;
         resolve();
       });
+      this.launchReject = err => {
+        proxy.exit().catch(this.log.debug);
+        reject(err);
+      };
     });
     return this.createExpander(proxy, builtinWorkspace, maxWorkspace);
   }
@@ -106,6 +120,8 @@ export abstract class AbstractLauncher implements ILanguageServerLauncher {
    */
   protected startupTimeout = 10000;
 
+  protected maxRespawn = 3;
+
   /**
    * try reconnect the proxy when disconnected
    */
@@ -122,19 +138,26 @@ export abstract class AbstractLauncher implements ILanguageServerLauncher {
       if (child && !child.killed && !processExpired()) {
         this.startConnect(proxy);
       } else {
-        if (child && this.running) {
-          this.log.debug('killing the old process.');
-          await this.killProcess(child);
+        if (this.spawnTimes < this.maxRespawn) {
+          if (child && this.running) {
+            this.log.debug('killing the old process.');
+            await this.killProcess(child);
+          }
+          const port = await this.getPort();
+          proxy.changePort(port);
+          this.child = await this.spawnProcess(installationPath, port, this.log);
+          this.spawnTimes += 1;
+          this.log.debug('spawned a child process ' + this.child.pid);
+          this._currentPid = this.child.pid;
+          this._startTime = Date.now();
+          this.running = true;
+          this.onProcessExit(this.child, () => this.reconnect(proxy, installationPath, child));
+          this.startConnect(proxy);
+        } else {
+          this.log.warn(`spawned process ${this.spawnTimes} times, mark this proxy unusable.`);
+          proxy.setError(ServerStartFailed);
+          this.launchReject!(ServerStartFailed);
         }
-        const port = await this.getPort();
-        proxy.changePort(port);
-        this.child = await this.spawnProcess(installationPath, port, this.log);
-        this.log.debug('spawned a child process ' + this.child.pid);
-        this._currentPid = this.child.pid;
-        this._startTime = Date.now();
-        this.running = true;
-        this.onProcessExit(this.child, () => this.reconnect(proxy, installationPath, child));
-        this.startConnect(proxy);
       }
     }
   }
@@ -142,7 +165,7 @@ export abstract class AbstractLauncher implements ILanguageServerLauncher {
   abstract async getPort(): Promise<number>;
 
   startConnect(proxy: LanguageServerProxy) {
-    proxy.connect();
+    proxy.connect().catch(this.log.debug);
   }
 
   /**
